update unit tests and fix bugs
